kvm tools: Seperate virtio-pci layer
This patch builds a virtio-pci layer which can be used by virtio
devices as a layer to interact with virtio-pci.

The purpose of the patch is to seperate the common virtio-pci layer from
being replicated in all virtio devices.

The new layer provides a callback interface to receive information about
virtio events.

Signed-off-by: Sasha Levin <levinsasha928@gmail.com>